+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Relation archetypes are removed when they are empty *and* the target entity is dead (#238, #242)
 * Support an unlimited number of cached filters, instead of 128 (#245)
+* `WorldStats` contains the number of cached filters (#246)
 
 ### Performance